# backend/ordersystem/src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Order Processing API
  version: v1
  description: >-
    Create, retrieve, list, advance, and cancel orders. The background pending-order
    processor has no public endpoint. Start the local service with ./dev to load
    stable recruiter fixtures for every lifecycle state and keep scheduling disabled.
servers:
  - url: /
    description: Current application origin.
security: []
tags:
  - name: Orders
    description: Create, find, browse, advance, and cancel orders in the V1 lifecycle.
paths:
  /api/v1/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create a new order
      description: |
        **Use this when:** a client is ready to save a new order.

        **What it does:** accepts 1–100 unique products and their quantities. The
        server assigns the order ID and timestamps and starts the order in `PENDING`.

        **What to expect:** success returns `201 Created`, the complete saved order,
        and a `Location` header containing its URL. Missing, invalid, or duplicate
        item data returns `400`. Sending the same request again creates another order;
        V1 does not support idempotency keys.
      requestBody:
        required: true
        description: Products and quantities to include in the new order.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrderRequest'
      responses:
        '201':
          description: Complete saved order in PENDING status.
          headers:
            Location:
              description: Relative URL of the created order.
              required: true
              schema:
                type: string
                format: uri-reference
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              examples:
                createdOrder:
                  $ref: '#/components/examples/PendingOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Orders]
      operationId: listOrders
      summary: Browse and filter orders
      description: |
        **Use this when:** a client needs an order history, queue, or status-filtered
        view without downloading every order's item details.

        **What it does:** returns one page of lightweight order summaries. Results are
        newest first. Optionally filter by one exact status and choose the page and
        page size.

        **What to expect:** success returns `200 OK`, including an empty `content`
        list when nothing matches. Pages start at `0`; the default size is `20` and
        the maximum is `100`. Invalid values or repeated documented parameters return
        `400`; unknown query parameters are ignored.
      parameters:
        - name: status
          in: query
          required: false
          description: Show only orders in this exact status. Leave blank to include every status.
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          required: false
          description: Page to return. Pages start at 0; the default is 0.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Maximum summaries to return. The default is 20 and the maximum is 100.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Requested page of order summaries and pagination totals.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResponse'
              examples:
                demoOrders:
                  $ref: '#/components/examples/DemoOrderPage'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Find one order by ID
      description: |
        **Use this when:** a client has an order ID and needs the complete current
        order, including all products, quantities, status, and timestamps.

        **What it does:** looks up the server-generated ID returned by the create or
        list endpoint. This is a read-only operation and does not change the order.

        **What to expect:** success returns `200 OK` with the complete order. An ID
        that is not a UUID returns `400`; a valid UUID that does not exist returns
        `404`.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Complete current order, including all items and timestamps.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              examples:
                pendingOrder:
                  $ref: '#/components/examples/PendingOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/status:
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Move an order to its next status
      description: |
        **Use this when:** fulfilment has completed the next step for an existing
        order.

        **What it does:** moves the order forward by exactly one step:
        `PENDING → PROCESSING → SHIPPED → DELIVERED`. Send only the next status in
        the request body. Use the separate cancel endpoint to cancel an order.

        **What to expect:** success returns `200 OK` with the updated complete order.
        Skipping a step, moving backward, repeating the current status, changing a
        terminal order, or losing a concurrent update returns `409`. Invalid input
        returns `400`. For a valid forward target (`PROCESSING`, `SHIPPED`, or
        `DELIVERED`), an unknown order returns `404`.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        description: The single next status that the order should enter.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Complete updated order after the status change.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              examples:
                processingOrder:
                  $ref: '#/components/examples/ProcessingOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderStateConflict'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel an order before processing
      description: |
        **Use this when:** a `PENDING` order must be stopped before processing starts.

        **What it does:** changes the order status to `CANCELLED`. It keeps the order
        and its history; cancellation never deletes data. Do not send a request body.

        **What to expect:** success returns `200 OK` with the complete cancelled order.
        An order that has started processing, is already cancelled, or loses a race
        with the background processor returns `409`. A malformed ID or any non-empty
        body—including `null`, `{}`, or whitespace—returns `400`; an unknown order
        returns `404`.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Complete updated order in CANCELLED status.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              examples:
                cancelledOrder:
                  $ref: '#/components/examples/CancelledOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderStateConflict'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order ID returned when the order was created, listed, or retrieved.
      schema:
        type: string
        format: uuid
      example: '11111111-1111-4111-8111-111111111111'
  headers:
    TraceId:
      description: Opaque server-generated request correlation identifier.
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    OrderStatus:
      type: string
      enum: [PENDING, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
    NewOrderRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          description: Product IDs must be exactly unique within this request.
          items:
            $ref: '#/components/schemas/NewOrderItemRequest'
      example:
        items:
          - productId: RECRUITER-HEADPHONES
            quantity: 2
          - productId: RECRUITER-KEYBOARD
            quantity: 1
    NewOrderItemRequest:
      type: object
      additionalProperties: false
      required: [productId, quantity]
      properties:
        productId:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^(?!\s*$)[^\u0000]+$'
          description: >-
            Opaque, case-sensitive identifier containing 1-100 Unicode code points;
            it must not be blank and must not contain U+0000.
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 999
    UpdateStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
      example:
        status: PROCESSING
    OrderItemResponse:
      type: object
      additionalProperties: false
      required: [productId, quantity]
      properties:
        productId:
          type: string
        quantity:
          type: integer
          format: int32
    OrderResponse:
      type: object
      additionalProperties: false
      required: [id, status, items, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderSummaryResponse:
      type: object
      additionalProperties: false
      required: [id, status, itemCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        itemCount:
          type: integer
          format: int64
          minimum: 1
          maximum: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderPageResponse:
      type: object
      additionalProperties: false
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummaryResponse'
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, traceId, violations]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - ORDER_NOT_FOUND
            - ORDER_STATE_CONFLICT
            - INTERNAL_ERROR
            - RESOURCE_NOT_FOUND
            - METHOD_NOT_ALLOWED
            - NOT_ACCEPTABLE
            - UNSUPPORTED_MEDIA_TYPE
        traceId:
          type: string
          minLength: 1
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
  examples:
    PendingOrder:
      summary: Seeded pending order
      value:
        id: '11111111-1111-4111-8111-111111111111'
        status: PENDING
        items:
          - productId: DEMO-PENDING-RED
            quantity: 2
          - productId: DEMO-PENDING-BLUE
            quantity: 1
        createdAt: '2026-07-15T10:00:00Z'
        updatedAt: '2026-07-15T10:00:00Z'
    ProcessingOrder:
      summary: Order after a legal status advance
      value:
        id: '11111111-1111-4111-8111-111111111111'
        status: PROCESSING
        items:
          - productId: DEMO-PENDING-RED
            quantity: 2
          - productId: DEMO-PENDING-BLUE
            quantity: 1
        createdAt: '2026-07-15T10:00:00Z'
        updatedAt: '2026-07-15T10:05:00Z'
    CancelledOrder:
      summary: Order after cancellation
      value:
        id: '11111111-1111-4111-8111-111111111111'
        status: CANCELLED
        items:
          - productId: DEMO-PENDING-RED
            quantity: 2
          - productId: DEMO-PENDING-BLUE
            quantity: 1
        createdAt: '2026-07-15T10:00:00Z'
        updatedAt: '2026-07-15T10:05:00Z'
    DemoOrderPage:
      summary: All five seeded lifecycle states
      value:
        content:
          - id: '55555555-5555-4555-8555-555555555555'
            status: CANCELLED
            itemCount: 1
            createdAt: '2026-07-15T10:04:00Z'
            updatedAt: '2026-07-15T10:05:00Z'
          - id: '44444444-4444-4444-8444-444444444444'
            status: DELIVERED
            itemCount: 1
            createdAt: '2026-07-15T10:03:00Z'
            updatedAt: '2026-07-15T10:05:00Z'
          - id: '33333333-3333-4333-8333-333333333333'
            status: SHIPPED
            itemCount: 1
            createdAt: '2026-07-15T10:02:00Z'
            updatedAt: '2026-07-15T10:04:00Z'
          - id: '22222222-2222-4222-8222-222222222222'
            status: PROCESSING
            itemCount: 1
            createdAt: '2026-07-15T10:01:00Z'
            updatedAt: '2026-07-15T10:02:00Z'
          - id: '11111111-1111-4111-8111-111111111111'
            status: PENDING
            itemCount: 2
            createdAt: '2026-07-15T10:00:00Z'
            updatedAt: '2026-07-15T10:00:00Z'
        page: 0
        size: 20
        totalElements: 5
        totalPages: 1
        first: true
        last: true
    InvalidRequestProblem:
      value:
        type: urn:problem:invalid-request
        title: Invalid request
        status: 400
        detail: The request is malformed or violates validation rules.
        instance: /api/v1/orders
        code: INVALID_REQUEST
        traceId: 8d4e9c1a
        violations: []
    OrderNotFoundProblem:
      value:
        type: urn:problem:order-not-found
        title: Order not found
        status: 404
        detail: The requested order does not exist.
        instance: /api/v1/orders/99999999-9999-4999-8999-999999999999
        code: ORDER_NOT_FOUND
        traceId: 8d4e9c1a
        violations: []
    OrderStateConflictProblem:
      value:
        type: urn:problem:order-state-conflict
        title: Order state conflict
        status: 409
        detail: The order cannot make the requested transition.
        instance: /api/v1/orders/55555555-5555-4555-8555-555555555555/cancel
        code: ORDER_STATE_CONFLICT
        traceId: 8d4e9c1a
        violations: []
    NotAcceptableProblem:
      value:
        type: urn:problem:not-acceptable
        title: Not acceptable
        status: 406
        detail: The requested response representation is not available.
        instance: /api/v1/orders
        code: NOT_ACCEPTABLE
        traceId: 8d4e9c1a
        violations: []
    UnsupportedMediaTypeProblem:
      value:
        type: urn:problem:unsupported-media-type
        title: Unsupported media type
        status: 415
        detail: The request media type is not supported.
        instance: /api/v1/orders
        code: UNSUPPORTED_MEDIA_TYPE
        traceId: 8d4e9c1a
        violations: []
    InternalErrorProblem:
      value:
        type: urn:problem:internal-error
        title: Internal error
        status: 500
        detail: An unexpected error occurred.
        instance: /api/v1/orders
        code: INTERNAL_ERROR
        traceId: 8d4e9c1a
        violations: []
  responses:
    InvalidRequest:
      description: The request is malformed or violates validation rules.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidRequest:
              $ref: '#/components/examples/InvalidRequestProblem'
    OrderNotFound:
      description: The requested order does not exist.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            orderNotFound:
              $ref: '#/components/examples/OrderNotFoundProblem'
    OrderStateConflict:
      description: The order cannot make the requested transition.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            orderStateConflict:
              $ref: '#/components/examples/OrderStateConflictProblem'
    NotAcceptable:
      description: The requested response representation is not available.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            notAcceptable:
              $ref: '#/components/examples/NotAcceptableProblem'
    UnsupportedMediaType:
      description: The request media type is not supported.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            unsupportedMediaType:
              $ref: '#/components/examples/UnsupportedMediaTypeProblem'
    InternalError:
      description: An unexpected error occurred.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            internalError:
              $ref: '#/components/examples/InternalErrorProblem'